requirements.txt
attrs==20.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
check50==3.3.11
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cs50==9.4.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
jellyfish==0.11.2
Jinja2==3.1.4
lib50==3.0.11
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
termcolor==1.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
gunicorn==20.1.0